% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{download_data_from_S3}
\alias{download_data_from_S3}
\title{download_data_from_S3}
\usage{
download_data_from_S3(
  data_local_path = "C:/Users/DIONE Oumar/Desktop/boston_data.csv",
  data_path_in_S3 = "Boston_House_price/Boston-house-price-data.csv",
  bucket_name = "odione",
  minio_object = my_minio
)
}
\arguments{
\item{minio_object}{}
}
\description{
download_data_from_S3
}
