% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_generation.R
\name{hitogramme_for_factor_var}
\alias{hitogramme_for_factor_var}
\title{hitogramme_for_factor_var}
\usage{
hitogramme_for_factor_var(
  df = boston_transformed_data,
  factor_col = "RAD",
  color_ = "blue"
)
}
\arguments{
\item{color_}{}
}
\description{
hitogramme_for_factor_var
}
