% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_generation.R
\name{plot_num_two_variable}
\alias{plot_num_two_variable}
\title{plot_num_two_variable}
\usage{
plot_num_two_variable(
  df,
  x_name,
  y_name,
  type_ = "p",
  title_ = "titre",
  color_ = "red"
)
}
\arguments{
\item{color_}{}
}
\description{
plot_num_two_variable
}
