% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_S3_connexion.R
\name{retrieve_minio}
\alias{retrieve_minio}
\title{retrieve_minio}
\usage{
retrieve_minio(
  file_path = "C:/Users/DIONE Oumar/Desktop/BostonDataPKG/configuration/config.yaml"
)
}
\arguments{
\item{file_path}{}
}
\description{
retrieve_minio
}
