% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{show_bucket_names}
\alias{show_bucket_names}
\title{show_bucket_names}
\usage{
show_bucket_names(minio = my_minio)
}
\arguments{
\item{minio}{an  S3 object for interacting with my bucket}
}
\value{
shows all my current S3 objects
}
\description{
show_bucket_names
}
