% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{upload_data_upon_S3}
\alias{upload_data_upon_S3}
\title{upload_data_upon_S3}
\usage{
upload_data_upon_S3(
  data_local_path = "C:/Users/DIONE Oumar/Desktop/boston_data.csv",
  data_path_in_S3 = "Boston_House_price/data_from_local.csv",
  bucket_name = "odione",
  minio_object = my_minio
)
}
\arguments{
\item{minio_object}{}
}
\description{
upload_data_upon_S3
}
